RedigerbarTekstRequest-klassa er eigentleg berre delvis til brevbakeren, vi bruker ho også til å finne rette greier i Gjenny-koden. Derfor misvisande med plasseringa som var, så flyttar ut
